moved buildModuleSubgraphSchema into lib/shared/util-graphql
