+++ apps/api-journeys/src/modules/application/index.ts
@@ -1,5 +1,5 @@
 import { createApplication } from 'graphql-modules'
-import { buildModuleSubgraphSchema } from './buildModuleSubgraphSchema'
+import { buildModuleSubgraphSchema } from '../../lib/shared/util-graphql/buildModuleSubgraphSchema'
 import journey from '../journey'
 
 export default createApplication({
